refactor: move logic to use-resume hook
