+++ public/locales/en/common.json
@@ -11,5 +11,6 @@
   "print-out": "Print Out",
   "add-new": "Add New",
   "your-role": "Your Role",
-  "your-education": "Your Education"
+  "your-education": "Your Education",
+  "select-template": "Select a Template"
 }
