+++ public/locales/pt/common.json
@@ -11,5 +11,6 @@
   "print-out": "Imprimir",
   "add-new": "Adicionar novo",
   "your-role": "Seu Cargo",
-  "your-education": "Sua Educação"
+  "your-education": "Sua Educação",
+  "select-template": "Selecione o modelo"
 }
