+++ src/components/templates/resume/index.ts
