+++ src/pages/_app.tsx
@@ -2,6 +2,8 @@ import type { AppProps } from 'next/app'
 import { appWithTranslation } from 'next-i18next'
 import { ThemeProvider } from 'styled-components'
 
+import { ResumeProvider } from 'hooks'
+
 import { DefaultSeo } from 'next-seo'
 import { defaultConfig } from '../../next-seo.config'
 
@@ -15,12 +17,14 @@ import 'antd/dist/antd.css'
 const MyApp = ({ Component, pageProps }: AppProps) => {
   return (
     <ThemeProvider theme={theme}>
-      <DefaultSeo {...defaultConfig} />
-      <GlobalStyles />
+      <ResumeProvider>
+        <DefaultSeo {...defaultConfig} />
+        <GlobalStyles />
 
-      <AppLayout>
-        <Component {...pageProps} />
-      </AppLayout>
+        <AppLayout>
+          <Component {...pageProps} />
+        </AppLayout>
+      </ResumeProvider>
     </ThemeProvider>
   )
 }
